+++ packages/salesforce/test/feature/index.test.js
@@ -1,11 +1,11 @@
 //--------------------------------------------------------
-//-- nwayo - Feature tests
+//-- Salesforce - Feature tests
 //--------------------------------------------------------
 import configTests           from '../../../../test/feature/config-tests';
 import { given, when, then } from './index.gwt';
 
 
-describe(`Validate that 'nwayo' works`, () => {
+describe(`Validate that 'salesforce' works`, () => {
 
 	configTests({ given, when, then });