+++ packages/salesforce/test/standards/index.test.js
@@ -1,0 +1,12 @@
+//--------------------------------------------------------
+//-- Salesforce - Standards tests
+//--------------------------------------------------------
+import packageTests          from '../../../../test/standards/package-tests';
+import { given, when, then } from './index.gwt';
+
+
+describe(`Validate 'salesforce' package.json`, () => {
+
+	packageTests({ given, when, then });
+
+});