+++ packages/scss/test/feature/index.test.js
@@ -1,0 +1,12 @@
+//--------------------------------------------------------
+//-- SCSS - Feature tests
+//--------------------------------------------------------
+import configTests           from '../../../../test/feature/config-tests';
+import { given, when, then } from './index.gwt';
+
+
+describe(`Validate that 'scss' works`, () => {
+
+	configTests({ given, when, then });
+
+});