+++ packages/scss/test/standards/index.test.js
@@ -1,11 +1,11 @@
 //--------------------------------------------------------
-//-- nwayo - Standards tests
+//-- SCSS - Standards tests
 //--------------------------------------------------------
 import packageTests          from '../../../../test/standards/package-tests';
 import { given, when, then } from './index.gwt';
 
 
-describe(`Validate 'nwayo' package.json`, () => {
+describe(`Validate 'scss' package.json`, () => {
 
 	packageTests({ given, when, then });